Refactor Twitter stream channel
