+++ lib/phoenix_app_web/channels/twitter_channel.ex
@@ -1,11 +1,11 @@
-defmodule PhoenixAppWeb.RoomChannel do
+defmodule PhoenixAppWeb.TwitterChannel do
   use Phoenix.Channel
 
-  def join("room:lobby", _message, socket) do
+  def join("twitter:stream", _message, socket) do
     { :ok, socket }
   end
 
-  def join("room:" <> _private_room_id, _params, _socket) do
+  def join("twitter:" <> _private_room_id, _params, _socket) do
     { :error, %{ reason: "unauthorized" } }
   end
